See description
Binaries are usually compiled and signed outside of GitHub. Since the
release would be a script that needs to be compiled using an
AppleScript compiler like Script Editor, it should be treated as a
binary. For releasing, I suggest making a GitHub Page for people to
download the compiled version and/or adding compiling instructions to
readme.md.